Add files via upload
Matlab Program for comparison of Images taken by the Camera.

This program helps to calculate the approximate lag between the cameras after synchronization. The test object can be a moving stick which can rapidly changes angle in a certain small duration. Both the camera take pictures of this angle change and then the angle is compared and plotted for different tests together by the program.